refact(stdx): move runtime functions into new package envx
